smartline-api/etl/old_etls/nfl-player-etl/requirements.txt
# NFL Player ETL Requirements
# Python 3.11+

# Database
psycopg2-binary==2.9.9

# HTTP (built-in http.client used, but requests can be alternative)
# requests==2.31.0  # Optional alternative to http.client

# Development/Testing (optional)
pytest==7.4.3
pytest-cov==4.1.0
python-dotenv==1.0.0
